+++ public_transport_watcher/monitoring/app.py
@@ -27,6 +27,6 @@ if st.session_state.logged_in:
         }
     )
 else:
-    pg = st.navigation([login_page, create_new_user_page])
+    pg = st.navigation([login_page])
 
 pg.run()